refine: left-align memoir, unify notes, cut experiments to five
Everything is left-aligned now — no centered text anywhere. The memoir
prose is roughly half its previous length.

Writing and experiments are combined into one system: notes, listed in
a new Appendix chapter on the homepage and rendered by a single
/notes/[slug] page that supports prose with embedded interactive
demos. The experiment list is cut from eighteen to five (dynamic
island, dock magnification, spring configurator, OKLCH, variable
font), each rewritten as a short essay. The WSL debugging post
migrates into the same system.

Old /experiments, /writing, and /photos routes redirect; their pages,
the 2,300-line experiments module, and ~40KB of dead CSS are removed.
