+++ app/page.tsx
@@ -5,11 +5,12 @@ import fs from "fs";
 import ThemeToggle from "@/components/ui/ThemeToggle";
 import Fade from "@/components/memoir/Fade";
 import PhotoPile from "@/components/memoir/PhotoPile";
+import { notes } from "@/app/notes/notes";
 
 export const metadata: Metadata = {
   title: "Tatenda Chinyamakobvu — Product Engineer",
   description:
-    "A short memoir of things built and things learned — where I come from, what I've made of it, and where I'm going. Product engineer in Harare, Zimbabwe.",
+    "A short memoir of things built and things learned. Product engineer in Harare, Zimbabwe.",
 };
 
 export default function Page() {
@@ -32,8 +33,8 @@ export default function Page() {
           <p className="memoir-role">Product Engineer · Harare, Zimbabwe</p>
           <hr className="memoir-rule" />
           <p className="memoir-lede">
-            This is less a portfolio than a short memoir — where I come from,
-            what I&apos;ve built, what I make of it, and where I&apos;m going.
+            Where I come from, what I&apos;ve built, what I make of it, and
+            where I&apos;m going.
           </p>
         </header>
       </Fade>
@@ -45,22 +46,18 @@ export default function Page() {
             <span className="memoir-kicker-title">Where I come from</span>
           </span>
           <p className="memoir-first">
-            I grew up in Zimbabwe, writing code for fun. There was no grand
-            plan — just a kid pulling things apart to see how they worked.
-            Fifteen years later I&apos;m still doing the same thing. The only
-            differences are that people now pay me for it, and the stakes are
-            a bit higher.
+            I grew up in Zimbabwe, writing code for fun. Fifteen years on
+            I&apos;m still that kid — people pay me now, and the stakes are
+            higher.
           </p>
           <p>
-            Zimbabwe teaches you things about software that no tutorial will.
-            The internet is negotiable. Power is negotiable. If your product
-            only works under perfect conditions, it doesn&apos;t work. That
-            constraint got into my taste early and never left: I build things
-            that keep working when the connection drops, because here it will.
+            Zimbabwe teaches you what no tutorial can. The internet is
+            negotiable. Power is negotiable. Software that only works in
+            perfect conditions doesn&apos;t work. That lesson got into my
+            taste early and never left.
           </p>
           <p>
-            The first years were the usual apprenticeship, compressed. I built
-            the buy-on-credit feature at{" "}
+            The apprenticeship came fast: a buy-on-credit feature at{" "}
             <a
               href="https://www.techzim.co.zw"
               target="_blank"
@@ -68,20 +65,16 @@ export default function Page() {
             >
               Techzim
             </a>
-            , did contract frontend work for{" "}
+            , contract work for{" "}
             <a
               href="https://www.22seven.com"
               target="_blank"
               rel="noopener noreferrer"
             >
               22seven
-            </a>{" "}
-            in South Africa, and led development of a multichain token bridge
-            at TreatDAO. I entered hackathons the way some people play sport —
-            first place at POTRAZ&apos;s Hack4SmartCities in 2020 for a
-            rapid-response safety app built on ambient sound, ninth of a
-            hundred and fifty teams at UmojaHack Africa the next year, a
-            bronze the year after that.
+            </a>
+            , a token bridge at TreatDAO. Hackathons the way other people play
+            sport — a first place, a ninth of a hundred and fifty, a bronze.
           </p>
         </section>
       </Fade>
@@ -90,9 +83,7 @@ export default function Page() {
         <section className="memoir-chapter">
           <span className="memoir-kicker">
             <span className="memoir-kicker-num">II</span>
-            <span className="memoir-kicker-title">
-              What I&apos;ve built, and what I think of it
-            </span>
+            <span className="memoir-kicker-title">What I&apos;ve built</span>
           </span>
           <p>
             The work I&apos;m proudest of is the least glamorous. Through{" "}
@@ -103,27 +94,16 @@ export default function Page() {
             >
               Corelith
             </a>
-            , my studio, three companies run their businesses on software I
-            built from scratch — a grocery retailer, a scrap-metal recycler,
-            and a mining operation. Point of sale, payroll, stock, financial
-            reporting, an append-only audit log. It loads its offline
-            workspace before it loads the dashboard, because it has to.
+            , three companies run their days on software I built from scratch
+            — a grocer, a scrap-metal recycler, a mine. Payroll, stock, an
+            audit log no one can quietly edit. It loads its offline workspace
+            before its dashboard, because here it has to.
           </p>
           <p>
-            What I&apos;ve come to believe is that businesses don&apos;t run
-            on clever code; they run on boring code that refuses to fall over.
-            The parts most engineers skip — sync, permissions, audit trails,
-            the CCTV camera grid that has to work from behind a NAT&apos;d
-            connection — are the parts a business actually feels. I{" "}
-            <a
-              href="https://github.com/tate2301/mediamtx-rtsp-forwarder"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              open-sourced
-            </a>{" "}
-            the small bridge that carries those on-prem camera feeds to the
-            browser, because someone else out there has the same problem.
+            Businesses don&apos;t run on clever code. They run on boring code
+            that refuses to fall over. The parts most engineers skip are the
+            parts a business actually feels — so those are the parts I build
+            first.
           </p>
           <p>
             At{" "}
@@ -134,36 +114,31 @@ export default function Page() {
             >
               Afrisight
             </a>{" "}
-            I was lead engineer on CXanalytics, their enterprise
-            customer-experience product. I drew the brand from scratch, built
-            the dashboard from a blank canvas, and wrote the targeting
-            algorithms that decide which survey reaches which respondent. I
-            also shipped the consumer mobile app that panelists across Africa
-            use daily — built so a survey never loses progress when a tower
-            drops. After I wrapped up full-time, they brought me back on
-            contract to build a no-code query builder over their raw survey
-            data: Stripe Sigma, applied to customer experience.
+            I built CXanalytics — the brand, the dashboard, the
+            survey-targeting algorithms — and the mobile app panelists across
+            Africa use daily, made so a survey survives a dropped tower. When
+            I left, they brought me back to build the query engine.
           </p>
           <p>
-            The rest is curiosity with a commit history.{" "}
+            The rest is curiosity with a commit history:{" "}
             <a
               href="https://github.com/tate2301/paynow-react"
               target="_blank"
               rel="noopener noreferrer"
             >
               paynow-react
-            </a>{" "}
-            became the standard React integration for Zimbabwe&apos;s dominant
-            payment gateway — never promoted, just useful. I built a{" "}
+            </a>
+            , quietly the standard React integration for Zimbabwe&apos;s
+            payment gateway, and a{" "}
             <a
               href="https://github.com/tate2301/paas-platform"
               target="_blank"
               rel="noopener noreferrer"
             >
-              self-hosted Heroku alternative
+              self-hosted Heroku
             </a>{" "}
-            not because the world needed another PaaS, but because the best
-            way to understand a system is to build it yourself.
+            built for one reason — you don&apos;t understand a system until
+            you&apos;ve built one.
           </p>
         </section>
       </Fade>
@@ -175,56 +150,76 @@ export default function Page() {
             <span className="memoir-kicker-title">Where I&apos;m going</span>
           </span>
           <p>
-            Corelith is the long project: quiet, dependable software that
-            Zimbabwean businesses run on — more tenants, more industries, the
-            same refusal to fall over. I want it to be the kind of company you
-            only hear about because everything works.
+            Corelith is the long project: software Zimbabwean businesses trust
+            with their day. I want to build the kind of company you only hear
+            about because everything works.
           </p>
           <p>
-            The other half is people. At the{" "}
+            The rest is people — student founders at the{" "}
             <a
               href="https://buildwithtif.xyz"
               target="_blank"
               rel="noopener noreferrer"
             >
               CUT Innovation Hub
-            </a>{" "}
-            I help student founders design and build products people actually
-            want to use, and I intend to keep making small, useful things in
-            public.
+            </a>
+            , and small useful things kept public.
           </p>
           <p>
-            I&apos;m available for work that&apos;s worth doing well. The
-            fastest way to reach me is{" "}
+            I&apos;m available for work worth doing well.{" "}
             <a href="mailto:thehalfstackdev@gmail.com">
               thehalfstackdev@gmail.com
             </a>
-            .
           </p>
         </section>
       </Fade>
 
       <Fade>
-        <section className="memoir-chapter memoir-photos">
+        <section className="memoir-chapter">
           <span className="memoir-kicker">
             <span className="memoir-kicker-num">IV</span>
             <span className="memoir-kicker-title">Photographs</span>
           </span>
           <p className="memoir-photos-intro">
-            Every archive should keep something that isn&apos;t work. These
-            are mine — favourite people, pups, and moments worth keeping.
+            Not everything worth keeping is work. Favourite people, pups,
+            moments.
           </p>
           <PhotoPile photos={photos} />
         </section>
       </Fade>
 
       <Fade>
+        <section className="memoir-chapter">
+          <span className="memoir-kicker">
+            <span className="memoir-kicker-num">V</span>
+            <span className="memoir-kicker-title">Appendix</span>
+          </span>
+          <p className="memoir-appendix-intro">
+            Notes — some written, some built. The built ones, you can touch.
+          </p>
+          <div className="appendix-list">
+            {notes.map((n) => (
+              <Link
+                key={n.slug}
+                href={`/notes/${n.slug}`}
+                className="appendix-row"
+              >
+                <span className="appendix-copy">
+                  <span className="appendix-title">{n.title}</span>
+                  <span className="appendix-line">{n.line}</span>
+                </span>
+                <span className="appendix-meta">
+                  {n.kind === "interactive" ? "interactive" : "written"} ·{" "}
+                  {n.date}
+                </span>
+              </Link>
+            ))}
+          </div>
+        </section>
+      </Fade>
+
+      <Fade>
         <footer className="memoir-foot">
-          <nav className="memoir-foot-row" aria-label="Site">
-            <Link href="/writing">writing</Link>
-            <Link href="/experiments">experiments</Link>
-            <Link href="/photos">photos</Link>
-          </nav>
           <div className="memoir-foot-row">
             <a
               href="https://github.com/tate2301"
