--- app/writing/(posts)/fixing-wsl-localhost-issues/page.mdx
@@ -1,23 +1,0 @@
-export const metadata = {
-  title: "Fixing WSL localhost & port forwarding issues — Tatenda Chinyamakobvu",
-  description:
-    "The smallest fix for the most annoying WSL2 networking bug — disable Fast Startup.",
-};
-
-# Fixing WSL localhost & port forwarding issues
-
-<p className="post-meta">October 22, 2025 · 1 min read</p>
-
-Firstly — man it has taken sooooo long to start writing anything on this website. Today was the spark. Let me set the scene.
-
-I'm working on a new project for a friend — it's a PaaS (Platform as a Service). Everything is going well, but suddenly I can't access my dev preview in the browser and VS Code becomes unresponsive. How? Well, I'm using WSL2 with Ubuntu. You probably know where this is going. A quick search on Google and you'll see how frustrating this issue is and how often it crops up. Unpredictable, everything just stops working.
-
-So here's the solution.
-
-**Disable Fast Startup.**
-
-That's it. Disable Fast Startup and restart. I don't know why, but it works. Cheers.
-
----
-
-[← back to writing](/writing)
