--- components/sections/Footer.tsx
@@ -1,13 +1,0 @@
-export default function Footer() {
-  return (
-    <footer>
-      <a href="mailto:thehalfstackdev@gmail.com">thehalfstackdev@gmail.com</a>
-      <span className="fsep" aria-hidden="true">·</span>
-      <a href="https://github.com/tate2301" target="_blank" rel="noopener noreferrer">GitHub</a>
-      <span className="fsep" aria-hidden="true">·</span>
-      <a href="https://www.upwork.com/freelancers/tatendachinyamakobvu" target="_blank" rel="noopener noreferrer">Upwork</a>
-      <span className="fsep" aria-hidden="true">·</span>
-      <a href="https://twitter.com/atipamara" target="_blank" rel="noopener noreferrer">X</a>
-    </footer>
-  );
-}
