--- content/experiments/article.mdx
@@ -1,3 +1,0 @@
-import { ExperimentDetailPage } from "@/components/sections/Experiments";
-
-<ExperimentDetailPage slug={props.slug} />
