--- content/experiments/resizable-panels.mdx
@@ -1,40 +1,0 @@
-import { ExperimentDemoBlock, ExperimentSourceBlock } from "@/components/sections/Experiments";
-import { Callout } from "@/components/mdx/Callout";
-
-# Resizable Panels
-
-<ExperimentDemoBlock id="resizable" />
-
-A two-panel split view where the divider can be dragged across the full width. The core mechanic is pointer capture — once the handle acquires the pointer, move events keep arriving even if the cursor leaves the element entirely.
-
-## Pointer capture
-
-Standard `mousemove` listeners break if the pointer moves faster than the event rate and leaves the handle element. `setPointerCapture` fixes this by routing all future events for that pointer ID to the capturing element, regardless of where the cursor actually is.
-
-```typescript
-const onDown = (e: React.PointerEvent) => {
-  e.currentTarget.setPointerCapture(e.pointerId);
-  dragging.current = true;
-};
-```
-
-On `pointerup` or `pointercancel`, capture is automatically released.
-
-<Callout type="insight">
-`pointercancel` is as important as `pointerup`. On mobile, a scroll gesture will cancel a pointer — without handling it, the drag state gets stuck active and the divider freezes until the next click.
-</Callout>
-
-## Snap to grid
-
-On release, the raw `split` percentage is rounded to the nearest 5% increment:
-
-```typescript
-const nearest = Math.round(split / 5) * 5;
-setSplit(nearest);
-```
-
-This makes it easy to hit common splits (50/50, 33/67, 25/75) without hunting. The transition back to the snap point is animated with `cubic-bezier(0.23, 1, 0.32, 1)` — a fast initial movement that decelerates, not a linear tween.
-
-The split is clamped to `[20%, 80%]` so neither panel collapses entirely.
-
-<ExperimentSourceBlock id="resizable" />
