--- content/experiments/view-transitions.mdx
@@ -1,51 +1,0 @@
-import { ExperimentDemoBlock, ExperimentSourceBlock } from "@/components/sections/Experiments";
-import { Callout } from "@/components/mdx/Callout";
-
-# View Transitions
-
-<ExperimentDemoBlock id="view-transition" />
-
-The View Transitions API lets the browser animate between two DOM states with a screenshot-and-crossfade mechanism. This demo drives a list-to-detail navigation: clicking a row transitions the entire panel, with named elements morphing between states.
-
-## How the API works
-
-```typescript
-document.startViewTransition(() => {
-  // Mutate DOM state here
-  setState(newState);
-});
-```
-
-The browser takes a screenshot before the callback runs, then a second screenshot after. It renders both as pseudo-elements — `::view-transition-old(root)` and `::view-transition-new(root)` — and plays a crossfade between them. The duration and easing are fully customizable with CSS.
-
-<Callout type="insight">
-`startViewTransition` wraps a synchronous state mutation, not an async operation. If your state update is synchronous React `setState`, it fires and React flushes it before the browser captures the "after" screenshot — no extra coordination needed.
-</Callout>
-
-## Named transitions
-
-Elements can opt into independent animations via `view-transition-name`. When both the old and new DOM have an element with the same name, the browser morphs between them with a shared-element transition rather than a full-page crossfade.
-
-```css
-/* The element moves and resizes between states */
-.list-item { view-transition-name: vt-item-1; }
-.detail-view { view-transition-name: vt-item-1; }
-```
-
-The demo assigns names dynamically based on item id, so each row has its own named transition.
-
-## Progressive enhancement
-
-The API is wrapped in a feature check:
-
-```typescript
-if ("startViewTransition" in document) {
-  document.startViewTransition(update);
-} else {
-  update();
-}
-```
-
-Browsers without the API still navigate — they just don't animate. No polyfill needed.
-
-<ExperimentSourceBlock id="view-transition" />
