+++ content/notes/the-day-localhost-lied.mdx
@@ -1,0 +1,9 @@
+It took me embarrassingly long to start writing anything here. This is what finally broke the silence.
+
+I was building a PaaS for a friend. Everything humming along — then the dev preview stops loading, VS Code hangs, and localhost is simply not there anymore. I'm on WSL2 with Ubuntu, so you probably know where this is going. Search the issue and you'll find years of people rediscovering the same misery: unpredictable, and everything just stops.
+
+Here is the entire fix.
+
+*Disable Fast Startup.* Restart Windows.
+
+That's it. I don't know why it works. I have made my peace with that. Cheers.
