+++ content/notes/the-dock-is-a-bell-curve.mdx
@@ -1,0 +1,19 @@
+import { Demo, Source } from "@/components/notes/demos";
+
+Everyone has run a cursor along the macOS Dock. Almost nobody can say why it feels like pushing a finger through water. I couldn't either, so I rebuilt it.
+
+<Demo id="dock" />
+
+The secret is that nothing zooms. Every icon's scale is a point on a Gaussian curve centred on your cursor:
+
+```
+scale(i) = 1 + (MAX − 1) × e^(−distance² / σ²)
+```
+
+There is no hovered icon. There is a field, and every icon lives in it — the one under your cursor peaks, and its neighbours rise with it in proportion.
+
+Sigma is the whole personality. At 30 the falloff is so steep it looks like a cheap hover effect. At 100 the entire dock inflates at once. At 58 it becomes a material.
+
+One more detail nobody notices: icons scale from their bottom edge, then get lifted so they never sink through the shelf. Delete that one line and the illusion collapses. That line is the difference between animation and physics.
+
+<Source id="dock" />
