+++ content/notes/why-your-gradients-turn-grey.mdx
@@ -1,0 +1,13 @@
+import { Demo, Source } from "@/components/notes/demos";
+
+Drag the two hues to opposite sides of the wheel — purple and green, blue and orange — and watch the top gradient die in the middle.
+
+<Demo id="oklch" />
+
+Both rows are given the same endpoints and the same request. The top row is sRGB, the way the web has mixed color for thirty years. The bottom is OKLCH. Only one of them keeps its promise.
+
+sRGB fails because it is an encoding, not a perception. Your eye does not weigh all hues equally, so a straight line through RGB space passes through colors that are mathematically halfway and visually dim. OKLCH was built the other way around: equal distances are supposed to look equal. Hold lightness and chroma steady, rotate the hue, and every step stays as vivid as the endpoints.
+
+This is why Tailwind, Radix, and Linear now generate their palettes in perceptual space — a grey-50 to grey-900 ramp should feel evenly spaced, and in sRGB it simply isn't. `oklch()` has shipped in every modern browser since 2023. There is no longer any excuse for a grey middle.
+
+<Source id="oklch" />
