--- posts/fixing-wsl-localhost-issues.mdx
@@ -1,11 +1,0 @@
-title: Fixing WSL localhost & port forwarding issues
-
-Firstly, man it has taken sooooo long to start writing anything on this website. Today was the spark actually. So let me set the scene.
-
-I'm working on a new project for a friend, its a PaaS(Platform as a Service). Everything is going well, but suddenly I can't access my dev preview in the browser and vscode becomes unresponsive. How? Well I am using WSL2 with Ubuntu. Guess you already know where this is going on. Just a quick search on google, you'll know how frustrating this issue is and how often it crops up. Unpredictable, everything just stops working.
-
-So here's the solution.
-
-*Disable Fast Startup!*
-
-That's it. Disable Fast Startup and restart. I don't know why, but it works. Cheers.
